astrology_db/requirements.txt
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.3
torch>=2.0.1
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.1
yt-dlp>=2023.3.4
deep-translator>=1.9.1
tqdm>=4.65.0
flask>=2.3.3
requests>=2.31.0
streamlit>=1.32.0
stripe==7.11.0 